requirements.txt
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.10.4
certifi==2023.11.17
distro==1.9.0
h11==0.14.0
httpcore==1.0.2
httpx==0.28.1
idna==3.6
joblib==1.4.0
langdetect==1.0.9
numpy==1.26.4
openai==1.25.0
pydantic==2.7.1
pydantic_core==2.18.2
python-telegram-bot==21.10
pytz==2024.1
PyYAML==6.0.3
scikit-learn==1.3.1
scipy==1.13.0
six==1.16.0
sniffio==1.3.0
threadpoolctl==3.5.0
tqdm==4.66.2
typing_extensions==4.11.0
tzlocal==5.2
